requirements.txt
kafka-python==2.0.2
requests-html==0.10.0
